requirement.txt
absl-py==1.4.0
accelerate==0.19.0
addict==2.4.0
albumentations==1.3.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
attrs==23.1.0
audioread==2.1.9
backcall==0.2.0
backports-zoneinfo==0.2.1
basicsr==1.3.5
braceexpand==0.1.7
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
clip==1.0
clip-anytorch==2.5.2
coca-pytorch==0.0.7
colorama==0.4.6
configargparse==1.5.3
coverage==4.0.3
cpbd==1.0.7
cycler==0.11.0
cython==0.29.30
dalle2-pytorch==1.14.2
dataclasses==0.6
decorator==4.4.2
dill==0.3.4
dlib==19.24.0
docopt==0.6.2
dominate==2.7.0
einops==0.6.1
ema-pytorch==0.2.3
embedding-reader==1.5.0
entrypoints==0.4
face-alignment==1.3.5
ffmpeg-python==0.2.0
filelock==3.11.0
flatbuffers==2.0.7
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.2.2
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.1.0
huggingface-hub==0.13.3
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
interval==1.0.0
iopath==0.1.10
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.3
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.11.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.6.8
libclang==16.0.0
librosa==0.7.2
llvmlite==0.31.0
lmdb==1.3.0
lpips==0.1.4
lws==1.2.7
mag==0.0.37
maggot==0.2
markdown==3.4.1
matplotlib==3.3.4
mediapipe==0.9.0.1
mmdnn==0.3.1
more-itertools==9.1.0
moviepy==1.0.3
mutagen==1.45.1
natsort==8.2.0
nest-asyncio==1.5.6
networkx==2.5.1
ninja==1.11.1
numba==0.48.0
numpy==1.21.6
nvdiffrast==0.3.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.10.0
open-clip-torch==2.20.0
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opentsne==0.6.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
pip==23.0.1
piq==0.7.0
pooch==1.6.0
portalocker==2.7.0
praat-parselmouth==0.4.3
proglog==0.1.9
prompt-toolkit==3.0.33
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
pygments==2.13.0
pynormalize==0.1.4
pyparsing==3.0.9
pyquaternion==0.9.9
pysptk==0.1.21
python-dateutil==2.8.1
python-speech-features==0.6
pytorch-fid==0.1.1
pytorch-warmup==0.1.1
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
pyvista==0.33.3
pywavelets==1.1.1
pyworld==0.3.0
pyyaml==6.0
pyzmq==19.0.2
qudida==0.0.4
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.3.1
resemblyzer==0.1.1.dev0
resize-right==0.0.2
rotary-embedding-torch==0.2.3
rsa==4.9
safetensors==0.3.1
scenedetect==0.6.0.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scooby==0.5.11
seaborn==0.11.2
sentencepiece==0.1.99
setuptools==58.0.4
six==1.15.0
soundfile==0.10.3.post1
sox==1.4.1
sync-batchnorm==0.0.1
tabulate==0.8.10
tb-nightly==2.12.0a20230113
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gpu==1.15.2
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.9.2
tokenizers==0.13.2
tomli==2.0.1
torch==1.8.0+cu111
torch-fidelity==0.3.0
torchaudio==0.8.0
torchfile==0.1.0
torchmetrics==0.11.4
torchvision==0.9.0+cu111
tornado==6.2
tqdm==4.65.0
traitlets==4.3.3
transformers==4.27.4
trimesh==3.20.2
typing==3.7.4.3
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.10
vector-quantize-pytorch==1.6.24
visdom==0.1.8.9
vtk==9.2.6
wcwidth==0.2.5
webdataset==0.2.48
webrtcvad==2.0.10
websocket-client==0.54.0
werkzeug==1.0.1
wget==3.2
wheel==0.37.1
wrapt==1.15.0
x-clip==0.12.1
xlwt==1.3.0
yacs==0.1.8
yapf==0.33.0
zipp==3.6.0